# [FlexAttention] Enforce Q,K,V memory layouts for fp8 flex attention to avoid perf degradation