@danielvegamyhre 2025-05-12
>I am wishy washy on if we want to does this under the hood a lil magically...

@drisspg I am as well, we could just add validation and require the user to do the transpose themselves, so they are explicitly aware of the layout of their tensors and there's no surprises? It does add some friction to using the API, though.

>That being said this fix shouldn't be needed on sm100 and newer which has TT support, can you make sure to keep this change to sm 90 devices

Good to know - updated.