@drisspg 2025-05-14
Im fine with enforcing for now, your perf benchmark above includes the transpose right?