@jeanschmidt 2025-05-15
@pytorchbot revert -m "Might have introduced regressions in rocm testing for main: https://github.com/pytorch/pytorch/actions/runs/15035410497/job/42257000513 feel free to re-merge if this was a mistake" -c nosignal