@danielvegamyhre 2025-05-15
Ergh, looks like graph will be different for AMD and CUDA now. Need to update the test