@danielvegamyhre 2025-05-15
Updated test to handle the different graph module produced for AMD.