@danielvegamyhre 2025-05-15
inductor test failing is related, error looks like a transient build/infra issue: 
```
[ RUN      ] AotInductorTest.BasicPackageLoaderTestCpu
unknown file: Failure
C++ exception with description "Error in dlopen: /usr/lib/x86_64-linux-gnu/libstdc++.so.6: version `GLIBCXX_3.4.30' not found (required by /tmp/j0sT4j/data/aotinductor/model/ctl2eqddy7dqff6vahjctdrmngtsn77afchsa3gojr2xfxyyysfb.wrapper.so)
```
rocm test failure is unrelated, error is transient test infra issue (test:
```
Run # copy test results back to the mounted workspace, needed sudo, resulting permissions were correct
Error: No such container: 
```